// web/src/app/user/api/type.ts
export interface PostCardProps {
    postId: string;
    userName: string;
    relation: string;
    avatar: string;
    time: string;
    title: string;
    content: string;
    images: string[];
    hashtags: string[];
    videos: string[];
    friends_tagged?: string[];
    feels: Feel[];
    feel_count: { [key: string]: number };
    comments: number;
    privacy: 'public' | 'friend';
    views: number;
    createdAt: string;
    post_link_meta?: PostLinkMeta
    handleGetPost: () => void;
    my_feel?: string; 
}
// ...
export interface PostLinkMeta {
    post_link_url?: string;
    post_link_title?: string;
    post_link_description?: string;
    post_link_content?: string;
     post_link_image?: string;
     post_link_video?: string;
     type?: string;
}